Basic raytracing works

+++ src/App/AppStateBase.h
@@ -16,8 +16,8 @@ namespace Raycer
 		virtual void resume() = 0;
 		virtual void shutdown() = 0;
 
-		virtual void update(double timeStep) = 0;
-		virtual void render(double timeStep, double interpolation) = 0;
+		virtual void update(float timeStep) = 0;
+		virtual void render(float timeStep, float interpolation) = 0;
 
 		virtual void windowResized(int width, int height) = 0;
 	};
